refactor(security): extract bounded mcporter registry read to own module
Keeps src/security/audit.ts under the TS max-LOC ceiling after the
bounded-read security fix.
